Silence warning

+++ models/eventmodel.cpp
@@ -21,7 +21,7 @@ int EventModel::rowCount(const QModelIndex& parent) const
     if (parent.isValid())
         return 0;
 
-    return eventList.size();
+    return static_cast<int>(eventList.size());
 }
 
 QVariant EventModel::data(const QModelIndex& index, int role) const
